Task #5077: Pull the trunk into the task branch

+++ branches/Cobalt-Task5077/RTCP/Cobalt/GPUProc/src/cuda/Kernels/BeamFormerTransposeKernel.h
@@ -42,7 +42,7 @@ namespace LOFAR
       struct Parameters : Kernel::Parameters
       {
         Parameters(const Parset& ps);
-        size_t nrTABs;
+        unsigned nrTABs;
       };
 
       enum BufferType
